@@ -37,6 +37,7 @@
   },
   "dependencies": {
     "@storybook/builder-webpack5": "7.0.0-beta.29",
+    "@storybook/cli": "7.0.0-beta.29",
     "@storybook/client-logger": "7.0.0-beta.29",
     "@storybook/core-client": "7.0.0-beta.29",
     "@storybook/core-common": "7.0.0-beta.29",
@@ -65,23 +66,22 @@
     "webpack": "5"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1303.5",
-    "@angular-devkit/build-angular": "^13.3.5",
-    "@angular-devkit/core": "^13.3.5",
-    "@angular/cli": "^13.3.5",
-    "@angular/common": "^13.3.6",
-    "@angular/compiler": "^13.3.6",
-    "@angular/compiler-cli": "^13.3.6",
-    "@angular/core": "^13.3.6",
-    "@angular/forms": "^13.3.6",
-    "@angular/platform-browser": "^13.3.6",
-    "@angular/platform-browser-dynamic": "^13.3.6",
-    "@nrwl/workspace": "14.6.1",
+    "@angular-devkit/architect": "^0.1500.4",
+    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular-devkit/core": "^15.0.4",
+    "@angular/cli": "^15.0.4",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/compiler-cli": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
     "@types/rimraf": "^3.0.2",
     "@types/tmp": "^0.2.3",
     "cross-spawn": "^7.0.3",
     "jest": "^29.3.1",
-    "jest-preset-angular": "^12.0.0",
+    "jest-preset-angular": "^12.2.3",
     "jest-specific-snapshot": "^7.0.0",
     "rimraf": "^3.0.2",
     "tmp": "^0.2.1",
@@ -90,19 +90,18 @@
     "zone.js": "^0.12.0"
   },
   "peerDependencies": {
-    "@angular-devkit/architect": ">=0.1300.0",
-    "@angular-devkit/build-angular": ">=13.0.0",
-    "@angular-devkit/core": ">=13.0.0",
-    "@angular/cli": ">=13.0.0",
-    "@angular/common": ">=13.0.0",
-    "@angular/compiler": ">=13.0.0",
-    "@angular/compiler-cli": ">=13.0.0",
-    "@angular/core": ">=13.0.0",
-    "@angular/forms": ">=13.0.0",
-    "@angular/platform-browser": ">=13.0.0",
-    "@angular/platform-browser-dynamic": ">=13.0.0",
+    "@angular-devkit/architect": ">=0.1400.0",
+    "@angular-devkit/build-angular": ">=14.0.0",
+    "@angular-devkit/core": ">=14.0.0",
+    "@angular/cli": ">=14.0.0",
+    "@angular/common": ">=14.0.0",
+    "@angular/compiler": ">=14.0.0",
+    "@angular/compiler-cli": ">=14.0.0",
+    "@angular/core": ">=14.0.0",
+    "@angular/forms": ">=14.0.0",
+    "@angular/platform-browser": ">=14.0.0",
+    "@angular/platform-browser-dynamic": ">=14.0.0",
     "@babel/core": "*",
-    "@nrwl/workspace": "14.6.1",
     "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
     "react-dom": "^16.8.0 || ^17.0.0 || ^18.0.0",
     "rxjs": "^6.0.0 || ^7.4.0",
@@ -111,9 +110,6 @@
   },
   "peerDependenciesMeta": {
     "@angular/cli": {
-      "optional": true
-    },
-    "@nrwl/workspace": {
       "optional": true
     }
   },
