@@ -1,9 +1,8 @@
-import AngularSnapshotSerializer from 'jest-preset-angular/build/AngularSnapshotSerializer';
-import HTMLCommentSerializer from 'jest-preset-angular/build/HTMLCommentSerializer';
+import AngularSnapshotSerializer from 'jest-preset-angular/build/serializers/ng-snapshot';
+import HTMLCommentSerializer from 'jest-preset-angular/build/serializers/html-comment';
 import { TestBed } from '@angular/core/testing';
-import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { addSerializer } from 'jest-specific-snapshot';
-import { getStorybookModuleMetadata } from '@storybook/angular/renderer';
+import { getApplication, storyPropsProvider } from '@storybook/angular/renderer';
 import { BehaviorSubject } from 'rxjs';
 
 addSerializer(HTMLCommentSerializer);
@@ -12,31 +11,20 @@ addSerializer(AngularSnapshotSerializer);
 function getRenderedTree(story: any) {
   const currentStory = story.render();
 
-  const moduleMeta = getStorybookModuleMetadata(
-    {
-      storyFnAngular: currentStory,
-      component: story.component,
-      // TODO : To change with the story Id in v7. Currently keep with static id to avoid changes in snapshots
-      targetSelector: 'storybook-wrapper',
-    },
-    new BehaviorSubject(currentStory.props)
-  );
-
-  TestBed.configureTestingModule({
-    imports: [...moduleMeta.imports],
-    declarations: [...moduleMeta.declarations],
-    providers: [...moduleMeta.providers],
-    schemas: [...moduleMeta.schemas],
+  const application = getApplication({
+    storyFnAngular: currentStory,
+    component: story.component,
+    // TODO : To change with the story Id in v7. Currently keep with static id to avoid changes in snapshots
+    targetSelector: 'storybook-wrapper',
   });
 
-  TestBed.overrideModule(BrowserDynamicTestingModule, {
-    set: {
-      entryComponents: [...moduleMeta.entryComponents],
-    },
+  TestBed.configureTestingModule({
+    imports: [application],
+    providers: [storyPropsProvider(new BehaviorSubject(currentStory.props))],
   });
 
   return TestBed.compileComponents().then(() => {
-    const tree = TestBed.createComponent(moduleMeta.bootstrap[0] as any);
+    const tree = TestBed.createComponent(application);
     tree.detectChanges();
 
     // Empty componentInstance remove attributes of the internal main component (<storybook-wrapper>) in snapshot
