@@ -11,13 +11,10 @@ import { CLIOptions } from '@storybook/types';
 import { catchError, map, mapTo, switchMap } from 'rxjs/operators';
 import { sync as findUpSync } from 'find-up';
 import { sync as readUpSync } from 'read-pkg-up';
-import {
-  BrowserBuilderOptions,
-  ExtraEntryPoint,
-  StylePreprocessorOptions,
-} from '@angular-devkit/build-angular';
+import { BrowserBuilderOptions, StylePreprocessorOptions } from '@angular-devkit/build-angular';
 
-import { buildStaticStandalone } from '@storybook/core-server';
+import { buildStaticStandalone, withTelemetry } from '@storybook/core-server';
+import { StyleElement } from '@angular-devkit/build-angular/src/builders/browser/schema';
 import { StandaloneOptions } from '../utils/standalone-options';
 import { runCompodoc } from '../utils/run-compodoc';
 import { buildStandaloneErrorHandler } from '../utils/build-standalone-errors-handler';
@@ -28,15 +25,17 @@ export type StorybookBuilderOptions = JsonObject & {
   docsMode: boolean;
   compodoc: boolean;
   compodocArgs: string[];
-  styles?: ExtraEntryPoint[];
+  styles?: StyleElement[];
   stylePreprocessorOptions?: StylePreprocessorOptions;
 } & Pick<
     // makes sure the option exists
     CLIOptions,
-    'outputDir' | 'configDir' | 'loglevel' | 'quiet' | 'webpackStatsJson'
+    'outputDir' | 'configDir' | 'loglevel' | 'quiet' | 'webpackStatsJson' | 'disableTelemetry'
   >;
 
 export type StorybookBuilderOutput = JsonObject & BuilderOutput & {};
+
+type StandaloneBuildOptions = StandaloneOptions & { outputDir: string };
 
 export default createBuilder<any, any>(commandBuilder);
 
@@ -65,15 +64,17 @@ function commandBuilder(
         outputDir,
         quiet,
         webpackStatsJson,
+        disableTelemetry,
       } = options;
 
-      const standaloneOptions: StandaloneOptions = {
+      const standaloneOptions: StandaloneBuildOptions = {
         packageJson: readUpSync({ cwd: __dirname }).packageJson,
         configDir,
         docsMode,
         loglevel,
         outputDir,
         quiet,
+        disableTelemetry,
         angularBrowserTarget: browserTarget,
         angularBuilderContext: context,
         angularBuilderOptions: {
@@ -83,6 +84,7 @@ function commandBuilder(
         tsConfig,
         webpackStatsJson,
       };
+
       return standaloneOptions;
     }),
     switchMap((standaloneOptions) => runInstance({ ...standaloneOptions, mode: 'static' })),
@@ -112,8 +114,15 @@ async function setup(options: StorybookBuilderOptions, context: BuilderContext) 
   };
 }
 
-function runInstance(options: StandaloneOptions) {
-  return from(buildStaticStandalone(options as any)).pipe(
-    catchError((error: any) => throwError(buildStandaloneErrorHandler(error)))
-  );
+function runInstance(options: StandaloneBuildOptions) {
+  return from(
+    withTelemetry(
+      'build',
+      {
+        cliOptions: options,
+        presetOptions: { ...options, corePresets: [], overridePresets: [] },
+      },
+      () => buildStaticStandalone(options)
+    )
+  ).pipe(catchError((error: any) => throwError(buildStandaloneErrorHandler(error))));
 }
