@@ -1,4 +1,5 @@
+export { storyPropsProvider } from './client/angular-beta/StorybookProvider';
 export { computesTemplateSourceFromComponent } from './client/angular-beta/ComputesTemplateFromComponent';
 export { rendererFactory } from './client/render';
 export { AbstractRenderer } from './client/angular-beta/AbstractRenderer';
-export { getStorybookModuleMetadata } from './client/angular-beta/StorybookModule';
+export { getApplication } from './client/angular-beta/StorybookModule';
