@@ -1,19 +1,13 @@
 import webpack from 'webpack';
 import { logger } from '@storybook/node-logger';
-import { BuilderContext, Target, targetFromTargetString } from '@angular-devkit/architect';
+import { BuilderContext, targetFromTargetString } from '@angular-devkit/architect';
 import { sync as findUpSync } from 'find-up';
-import semver from 'semver';
 import { dedent } from 'ts-dedent';
 
 import { JsonObject, logging } from '@angular-devkit/core';
 import { getWebpackConfig as getCustomWebpackConfig } from './angular-cli-webpack';
 import { moduleIsAvailable } from './utils/module-is-available';
 import { PresetOptions } from './preset-options';
-import {
-  getDefaultProjectName,
-  findAngularProjectTarget,
-  readAngularWorkspaceConfig,
-} from './angular-read-workspace';
 
 export async function webpackFinal(baseConfig: webpack.Configuration, options: PresetOptions) {
   if (!moduleIsAvailable('@angular-devkit/build-angular')) {
@@ -21,43 +15,18 @@ export async function webpackFinal(baseConfig: webpack.Configuration, options: P
     return baseConfig;
   }
 
-  const angularCliVersion = await import('@angular/cli').then((m) => semver.coerce(m.VERSION.full));
+  checkForLegacyBuildOptions(options);
 
-  /**
-   * Ordered array to use the specific  getWebpackConfig according to some condition like angular-cli version
-   */
-  const webpackGetterByVersions: {
-    info: string;
-    condition: boolean;
-    getWebpackConfig(
-      baseConfig: webpack.Configuration,
-      options: PresetOptions
-    ): Promise<webpack.Configuration> | webpack.Configuration;
-  }[] = [
-    {
-      info: '=> Loading angular-cli config for angular >= 13.0.0',
-      condition: semver.satisfies(angularCliVersion, '>=13.0.0'),
-      getWebpackConfig: async (_baseConfig, _options) => {
-        const builderContext = getBuilderContext(_options);
-        const builderOptions = await getBuilderOptions(_options, builderContext);
-        const legacyDefaultOptions = await getLegacyDefaultBuildOptions(_options);
+  const builderContext = getBuilderContext(options);
+  const builderOptions = await getBuilderOptions(options, builderContext);
 
-        return getCustomWebpackConfig(_baseConfig, {
-          builderOptions: {
-            watch: options.configType === 'DEVELOPMENT',
-            ...legacyDefaultOptions,
-            ...builderOptions,
-          },
-          builderContext,
-        });
-      },
+  return getCustomWebpackConfig(baseConfig, {
+    builderOptions: {
+      watch: options.configType === 'DEVELOPMENT',
+      ...builderOptions,
     },
-  ];
-
-  const webpackGetter = webpackGetterByVersions.find((wg) => wg.condition);
-
-  logger.info(webpackGetter.info);
-  return Promise.resolve(webpackGetter.getWebpackConfig(baseConfig, options));
+    builderContext,
+  });
 }
 
 /**
@@ -116,58 +85,23 @@ async function getBuilderOptions(
   return builderOptions;
 }
 
+export const migrationToBuilderReferrenceMessage = dedent`Your Storybook startup uses a solution that is not supported.
+    You must use angular builder to have an explicit configuration on the project used in angular.json
+    Read more at:
+    - https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#sb-angular-builder)
+    - https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#angular13)
+  `;
+
 /**
- * Get options from legacy way
- * /!\ This is only for backward compatibility and would be removed on Storybook 7.0
- * only work for angular.json with [defaultProject].build or "storybook.build" config
+ * Checks if using legacy configuration that doesn't use builder and logs message referring to migration docs.
  */
-async function getLegacyDefaultBuildOptions(options: PresetOptions) {
+function checkForLegacyBuildOptions(options: PresetOptions) {
   if (options.angularBrowserTarget !== undefined) {
     // Not use legacy way with builder (`angularBrowserTarget` is defined or null with builder and undefined without)
-    return {};
+    return;
   }
 
-  logger.warn(dedent`Your Storybook startup uses a solution that will not be supported in version 7.0. 
-            You must use angular builder to have an explicit configuration on the project used in angular.json
-            Read more at:
-            - https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#sb-angular-builder)
-            - https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#angular13)
-          `);
-  const dirToSearch = process.cwd();
+  logger.error(migrationToBuilderReferrenceMessage);
 
-  // Read angular workspace
-  let workspaceConfig;
-  try {
-    workspaceConfig = await readAngularWorkspaceConfig(dirToSearch);
-  } catch (error) {
-    logger.error(
-      `=> Could not find angular workspace config (angular.json) on this path "${dirToSearch}"`
-    );
-    logger.info(`=> Fail to load angular-cli config. Using base config`);
-    return {};
-  }
-
-  // Find angular project target
-  try {
-    const browserTarget = {
-      configuration: undefined,
-      project: getDefaultProjectName(workspaceConfig),
-      target: 'build',
-    } as Target;
-
-    const { target } = findAngularProjectTarget(
-      workspaceConfig,
-      browserTarget.project,
-      browserTarget.target
-    );
-
-    logger.info(
-      `=> Using angular project "${browserTarget.project}:${browserTarget.target}" for configuring Storybook`
-    );
-    return { ...target.options };
-  } catch (error) {
-    logger.error(`=> Could not find angular project: ${error.message}`);
-    logger.info(`=> Fail to load angular-cli config. Using base config`);
-    return {};
-  }
+  throw Error('angularBrowserTarget is undefined.');
 }
