@@ -17,13 +17,16 @@ const defaultOptions: FrameworkOptions = {
   extraAddons: [],
   staticDir: undefined,
   addScripts: true,
+  addMainFile: true,
   addComponents: true,
   addBabel: false,
   addESLint: false,
   extraMain: undefined,
   framework: undefined,
   extensions: undefined,
+  componentsDestinationPath: undefined,
   commonJs: false,
+  storybookConfigFolder: '.storybook',
 };
 
 const getBuilderDetails = (builder: string) => {
@@ -106,7 +109,13 @@ const hasFrameworkTemplates = (framework?: SupportedFrameworks) =>
 export async function baseGenerator(
   packageManager: JsPackageManager,
   npmOptions: NpmOptions,
-  { language, builder = CoreBuilder.Webpack5, pnp, commonJs }: GeneratorOptions,
+  {
+    language,
+    builder = CoreBuilder.Webpack5,
+    pnp,
+    commonJs,
+    frameworkPreviewParts,
+  }: GeneratorOptions,
   renderer: SupportedRenderers,
   options: FrameworkOptions = defaultOptions,
   framework?: SupportedFrameworks
@@ -116,11 +125,14 @@ export async function baseGenerator(
     extraPackages,
     staticDir,
     addScripts,
+    addMainFile,
     addComponents,
     addBabel,
     addESLint,
     extraMain,
     extensions,
+    storybookConfigFolder,
+    componentsDestinationPath,
   } = {
     ...defaultOptions,
     ...options,
@@ -194,26 +206,29 @@ export async function baseGenerator(
 
   const versionedPackages = await packageManager.getVersionedPackages(packages);
 
-  await fse.ensureDir('./.storybook');
+  await fse.ensureDir(`./${storybookConfigFolder}`);
 
-  await configureMain({
-    framework: { name: frameworkInclude, options: options.framework || {} },
-    docs: { autodocs: 'tag' },
-    addons: pnp ? addons.map(wrapForPnp) : addons,
-    extensions,
-    commonJs,
-    ...(staticDir ? { staticDirs: [path.join('..', staticDir)] } : null),
-    ...extraMain,
-    ...(type !== 'framework'
-      ? {
-          core: {
-            builder: builderInclude,
-          },
-        }
-      : {}),
-  });
+  if (addMainFile) {
+    await configureMain({
+      framework: { name: frameworkInclude, options: options.framework || {} },
+      storybookConfigFolder,
+      docs: { autodocs: 'tag' },
+      addons: pnp ? addons.map(wrapForPnp) : addons,
+      extensions,
+      commonJs,
+      ...(staticDir ? { staticDirs: [path.join('..', staticDir)] } : null),
+      ...extraMain,
+      ...(type !== 'framework'
+        ? {
+            core: {
+              builder: builderInclude,
+            },
+          }
+        : {}),
+    });
+  }
 
-  await configurePreview(rendererId);
+  await configurePreview({ frameworkPreviewParts, storybookConfigFolder });
 
   // FIXME: temporary workaround for https://github.com/storybookjs/storybook/issues/17516
   if (
@@ -226,7 +241,9 @@ export async function baseGenerator(
         window.global = window;
       </script>
     `;
-    await fse.writeFile(`.storybook/preview-head.html`, previewHead, { encoding: 'utf8' });
+    await fse.writeFile(`${storybookConfigFolder}/preview-head.html`, previewHead, {
+      encoding: 'utf8',
+    });
   }
 
   const babelDependencies =
@@ -256,6 +273,6 @@ export async function baseGenerator(
 
   if (addComponents) {
     const templateLocation = hasFrameworkTemplates(framework) ? framework : rendererId;
-    await copyComponents(templateLocation, language);
+    await copyComponents(templateLocation, language, componentsDestinationPath);
   }
 }
