@@ -34,9 +34,11 @@ import { extractStorybookMetadata } from './utils/metadata';
 import { StoryIndexGenerator } from './utils/StoryIndexGenerator';
 import { summarizeIndex } from './utils/summarizeIndex';
 
-export async function buildStaticStandalone(
-  options: CLIOptions & LoadOptions & BuilderOptions & { outputDir: string }
-) {
+export type BuildStaticStandaloneOptions = CLIOptions &
+  LoadOptions &
+  BuilderOptions & { outputDir: string };
+
+export async function buildStaticStandalone(options: BuildStaticStandaloneOptions) {
   /* eslint-disable no-param-reassign */
   options.configType = 'PRODUCTION';
 
